Find and download refactor for protocol. Download and find functions work
