+++ gpm_api/io/info.py
@@ -115,7 +115,11 @@ def _get_info_from_filename(fname):
             raise ValueError(f"{fname} can not be parsed. Report the issue.")
 
     # Add product information
-    info_dict["product"] = get_product_from_filepath(fname)
+    # - ATTENTION: can not be inferred for products not defined in etc/product.yml
+    try:
+        info_dict["product"] = get_product_from_filepath(fname)
+    except Exception:
+        pass
 
     # Return info dictionary
     return info_dict
