+++ gpm_api/scripts/download_gpm_monthly_data.py
@@ -20,6 +20,7 @@ sys.tracebacklimit = 0  # avoid full traceback error if occur
 @click.argument("year", type=int)
 @click.argument("month", type=int)
 @click.option("--product_type", type=str, show_default=True, default="RS")
+@click.option("--protocol", type=str, show_default=True, default="pps")
 @click.option("--version", type=int, show_default=True, default=None)
 @click.option("--n_threads", type=int, default=4)
 @click.option("--transfer_tool", type=str, default="curl")
@@ -34,6 +35,7 @@ def download_gpm_monthly_data(
     year,
     month,
     product_type="RS",
+    protocol="pps",
     version=None,
     n_threads=4,
     transfer_tool="curl",
@@ -53,6 +55,7 @@ def download_gpm_monthly_data(
         month=month,
         product_type=product_type,
         version=version,
+        protocol=protocol,
         n_threads=n_threads,
         transfer_tool=transfer_tool,
         progress_bar=progress_bar,
