+++ gpm_api/tests/0_tmp/io/test_disk.py
@@ -6,7 +6,7 @@ Created on Fri Dec  9 15:46:47 2022
 @author: ghiggi
 """
 import datetime
-from gpm_api.io.disk import _find_daily_filepaths, find_filepaths
+from gpm_api.io.find import find_daily_filepaths, find_filepaths
 
 product = "2A-DPR"
 date = datetime.date(2020, 7, 5)  # OK
@@ -17,7 +17,8 @@ end_time = None
 product_type = "RS"
 verbose = True
 
-filepaths = _find_daily_filepaths(
+filepaths, _ = find_daily_filepaths(
+    protocol="local",
     product=product,
     product_type=product_type,
     date=date,
@@ -32,6 +33,7 @@ start_time = datetime.datetime(2020, 7, 5, 0, 2, 0)
 end_time = datetime.datetime(2020, 7, 5, 0, 3, 0)
 
 filepaths = find_filepaths(
+    protocol="local",
     product=product,
     product_type=product_type,
     version=version,
